+++ test21/app/src/main/res/drawable/example_resize.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list  xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:drawable="@drawable/abc_vector_test"
+        android:width="200dp"
+        android:height="200dp" />
+
+</layer-list>
+
