+++ test21/app/src/main/res/layout/activity_main.xml
@@ -7,55 +7,24 @@
     android:orientation="vertical"
     tools:context=".MainActivity">
 
-    <!--  EditText 설정
-    1. ems : 넓이 길이를  wrap_content 로 하면 1칸만 나오는데
-             EditText의  보여지는 입력칸 수를 설정가능
-               => 혹은 그냥 match_parent로 설정해도됨
-    2. android:inputType
-        none : 하단 버튼이 엔터. 글자를 쓰면 밑으로 내려감
-        text : 하단 모양이 확인으로 바뀜. 글자쓰면 옆으로 길게 이어지고 엔터처리안됨
-               확인 클릭시 키보드 닫힘
-        자주쓰틑 이메일, 전화번호, 나이, 등등 가능
-    3. 입력전 힌트임
-   -->
-    <EditText
-        android:id="@+id/et_email"
+<!--    1. android:background="@null" : 한 줄이 보이지 않음
+        2. android:maxLength="10" : 최대 길이 설정-->
+    <!--<EditText
+        android:id="@+id/et"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:inputType="textEmailAddress"
-        android:hint="이메일"
-     />
-    <EditText
-        android:id="@+id/et_password"
+        android:background="@null"
+        android:maxLength="10"
+        />-->
+
+    <!-- 이미지뷰 비율일정
+         android:src="@null "
+-->
+    <ImageView
+        android:src="@color/design_default_color_primary"
+        android:scaleType="center"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:inputType="textPassword"
-        android:hint="비밀번호"
-        />
-    <EditText
-        android:id="@+id/et_name"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:inputType="text"
-        android:hint="이름"
-        />
-    <EditText
-        android:id="@+id/et_age"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:inputType="number"
-        android:hint="나이"
-        />
-    <Button
-        android:onClick="Login"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="로그인"
-        android:layout_gravity="center_horizontal"
-        android:backgroundTint="#3B8FEA"
-        android:textColor="#fff"
-        android:textStyle="bold"
+        android:layout_height="match_parent"
 
         />
-
 </LinearLayout>
